@@ -1,15 +1,15 @@
 {
   "name": "todolist-frontend",
   "homepage": ".",
-  "version": "0.2.1",
+  "version": "1.0.0",
   "private": true,
   "dependencies": {
     "axios": "^0.21.1",
-    "node-sass": "^4.13.0",
-    "react": "^16.11.0",
+    "node-sass": "^6.0.1",
+    "react": "^17.0.2",
     "react-day-picker": "^7.4.0",
-    "react-dom": "^16.11.0",
-    "react-scripts": "3.2.0"
+    "react-dom": "^17.0.2",
+    "react-scripts": "4.0.3"
   },
   "scripts": {
     "start": "BROWSER=firefox react-scripts start",
@@ -34,6 +34,9 @@
     ]
   },
   "devDependencies": {
-    "eslint": "^6.6.0"
+    "eslint": "^7.32.0",
+    "eslint-config-prettier": "^8.3.0",
+    "eslint-plugin-prettier": "^4.0.0",
+    "prettier": "^2.4.1"
   }
 }
