@@ -1,27 +1,35 @@
-var express = require('express')
-import { createList, getAllLists, addListItem, deleteListItem, updateListItem, deleteList, findList } from '../controllers/listController'
-var router = express.Router()
+var express = require("express");
+import {
+  createList,
+  getAllLists,
+  addListItem,
+  deleteListItem,
+  updateListItem,
+  deleteList,
+  findList,
+} from "../controllers/listController";
+var router = express.Router();
 
-router.get('/', function(req, res, next) {
-  res.render('index', { title: 'JONAS' })
-})
-router.get('/lists', function(req, res, next) {
+router.get("/", function (req, res, next) {
+  res.render("index", { title: "JONAS" });
+});
+router.get("/lists", function (req, res, next) {
   getAllLists()
-    .then(lists => {
-      return res.json(lists)
+    .then((lists) => {
+      return res.json(lists);
     })
-    .catch(err => {
-      return res.status(500).send('Internal Server Error ' + err.message)
-    })
-})
-router.get('/getlist', function(req, res, next) {
-  findList({}).then(list => {
-    return res.json(list)
-  })
+    .catch((err) => {
+      return res.status(500).send("Internal Server Error " + err.message);
+    });
+});
+router.get("/getlist", function (req, res, next) {
+  findList({}).then((list) => {
+    return res.json(list);
+  });
   // .catch(err => {
   //   return res.json({ status: 'ERROR', message: err.message })
   // })
-})
+});
 
 // not used in demo
 // router.post('/createlist', function(req, res, next) {
@@ -36,62 +44,65 @@ router.get('/getlist', function(req, res, next) {
 // })
 
 //for testing purposes only
-// router.get('/createlist', function(req, res, next) {
-//   const { listTitle } = req.body
-//   createList({ title: listTitle })
-//     .then(list => {
-//       return res.status(200).json(list)
-//     })
-//     .catch(err => {
-//       return res.status(500).send('Internal Server Error ' + err.message)
-//     })
-// })
+router.get("/createlist", function (req, res, next) {
+  // const { listTitle } = req.body;
 
-router.post('/additem/:list_id', function(req, res, next) {
-  const { list_id } = req.params
-  const { data } = req.body
+  const listTitle = "lista";
+
+  createList({ title: listTitle })
+    .then((list) => {
+      return res.status(200).json(list);
+    })
+    .catch((err) => {
+      return res.status(500).send("Internal Server Error " + err.message);
+    });
+});
+
+router.post("/additem/:list_id", function (req, res, next) {
+  const { list_id } = req.params;
+  const { data } = req.body;
 
   addListItem({ list_id, data })
-    .then(item => {
-      return res.status(200).json(item)
+    .then((item) => {
+      return res.status(200).json(item);
     })
-    .catch(err => {
-      return res.status(500).send('Internal Server Error ' + err.message)
-    })
-})
+    .catch((err) => {
+      return res.status(500).send("Internal Server Error " + err.message);
+    });
+});
 
-router.put('/updateitem/:list_id/:item_id', function(req, res, next) {
-  const { list_id, item_id } = req.params
-  const { data } = req.body
+router.put("/updateitem/:list_id/:item_id", function (req, res, next) {
+  const { list_id, item_id } = req.params;
+  const { data } = req.body;
   updateListItem({ list_id, item_id, data })
     .then(() => {
-      return res.status(204).send()
+      return res.status(204).send();
     })
-    .catch(err => {
-      return res.status(500).send('Internal Server Error ' + err.message)
-    })
-})
-router.delete('/deleteitem/:list_id/:item_id', function(req, res, next) {
-  const { list_id, item_id } = req.params
+    .catch((err) => {
+      return res.status(500).send("Internal Server Error " + err.message);
+    });
+});
+router.delete("/deleteitem/:list_id/:item_id", function (req, res, next) {
+  const { list_id, item_id } = req.params;
 
   deleteListItem({ list_id, item_id })
     .then(() => {
-      return res.status(204).send()
+      return res.status(204).send();
     })
-    .catch(err => {
-      return res.status(500).send('Internal Server Error ' + err.message)
-    })
-})
+    .catch((err) => {
+      return res.status(500).send("Internal Server Error " + err.message);
+    });
+});
 
-router.delete('/deletelist/:list_id', function(req, res, next) {
-  const { list_id } = req.params
+router.delete("/deletelist/:list_id", function (req, res, next) {
+  const { list_id } = req.params;
   deleteList({ list_id })
-    .then(message => {
-      return res.json({ status: 'OK', message })
+    .then((message) => {
+      return res.json({ status: "OK", message });
     })
-    .catch(err => {
-      return res.json({ status: 'ERROR', message: err.message })
-    })
-})
+    .catch((err) => {
+      return res.json({ status: "ERROR", message: err.message });
+    });
+});
 
-export default router
+export default router;
