@@ -1,2 +1,5 @@
-require = require('esm')(module)
-module.exports = require('./bin/www')
+require = require("esm")(module);
+module.exports = require("./bin/www");
+
+// require = require("esm")(module/*, options*/)
+// module.exports = require("./main.js")
